neural networks/endpoint/backend/requirements.txt
numpy<2.0.0
torch==2.2.1
torchvision==0.17.1
fastapi==0.109.1
uvicorn==0.27.0
pillow==10.2.0
python-multipart==0.0.6